Sockets: separate api to its own source file
