+++ applications/services/sockets/application.fam
@@ -9,6 +9,7 @@ App(
     requires=["wifi"],
     sources=[
         "sockets.c",
+        "sockets_api.c",
         "sockets_common.c",
     ],
 )
